requirements.txt
pip
python
pygame
neat-python